Remove dependency of Array/Types on StreamK
It can depend on StreamK/Type but not on the high level StreamK module.
